+++ libraries/tuttle/src/tuttle/host/ofx/ofxhInteract.cpp
@@ -165,7 +165,7 @@ namespace Interact {
       }
       
       // do nothing
-      int Instance::getDimension(const std::string &name) const OFX_EXCEPTION_SPEC
+      size_t Instance::getDimension(const std::string &name) const OFX_EXCEPTION_SPEC
       {
         if(name == kOfxInteractPropPixelScale){
           return 2;
